Remove global resources e2e test
