+++ test/e2e/cluster_test.go
@@ -1,12 +1,7 @@
 package tests
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io"
-	"log"
-	"net/http"
 	"strings"
 	"time"
 
@@ -21,28 +16,9 @@ import (
 
 	"github.com/stolostron/multicluster-global-hub/pkg/constants"
 	"github.com/stolostron/multicluster-global-hub/pkg/database/models"
-	"github.com/stolostron/multicluster-global-hub/pkg/utils"
-)
-
-const (
-	CLUSTER_LABEL_KEY   = "label"
-	CLUSTER_LABEL_VALUE = "test"
 )
 
 var _ = Describe("Managed Clusters", Label("e2e-test-cluster"), Ordered, func() {
-	Context("Cluster Label", func() {
-		It("add cluster label", func() {
-			for _, managedCluster := range managedClusters {
-				assertAddLabel(managedCluster, CLUSTER_LABEL_KEY, CLUSTER_LABEL_VALUE)
-			}
-		})
-
-		It("remove cluster label", func() {
-			for _, managedCluster := range managedClusters {
-				assertRemoveLabel(managedCluster, CLUSTER_LABEL_KEY, CLUSTER_LABEL_VALUE)
-			}
-		})
-	})
 
 	Context("Cluster Events", func() {
 		It("sync the event to the global hub database", func() {
@@ -213,169 +189,6 @@ var _ = Describe("Managed Clusters", Label("e2e-test-cluster"), Ordered, func() 
 		})
 	})
 })
-
-type patch struct {
-	Op    string `json:"op" binding:"required"`
-	Path  string `json:"path" binding:"required"`
-	Value string `json:"value"`
-}
-
-func getManagedCluster(client *http.Client) ([]clusterv1.ManagedCluster, error) {
-	managedClusterUrl := fmt.Sprintf("%s/global-hub-api/v1/managedclusters", testOptions.GlobalHub.Nonk8sApiServer)
-	req, err := http.NewRequest("GET", managedClusterUrl, nil)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			log.Printf("failed to close response body: %v", err)
-		}
-	}()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	var clusterList clusterv1.ManagedClusterList
-	err = json.Unmarshal(body, &clusterList)
-	if err != nil {
-		return nil, err
-	}
-
-	return clusterList.Items, nil
-}
-
-func getManagedClusterByName(client *http.Client, managedClusterName string) (
-	*clusterv1.ManagedCluster, error,
-) {
-	managedClusterUrl := fmt.Sprintf("%s/global-hub-api/v1/managedclusters",
-		testOptions.GlobalHub.Nonk8sApiServer)
-	req, err := http.NewRequest("GET", managedClusterUrl, nil)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() {
-		if err := resp.Body.Close(); err != nil {
-			log.Printf("failed to close response body: %v", err)
-		}
-	}()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	var managedClusterList clusterv1.ManagedClusterList
-	err = json.Unmarshal(body, &managedClusterList)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, managedCluster := range managedClusterList.Items {
-		if managedCluster.Name == managedClusterName {
-			return &managedCluster, nil
-		}
-	}
-
-	return nil, nil
-}
-
-func assertAddLabel(cluster clusterv1.ManagedCluster, labelKey, labelVal string) {
-	patches := []patch{
-		{
-			Op:    "add",
-			Path:  "/metadata/labels/" + labelKey,
-			Value: labelVal,
-		},
-	}
-
-	By("Check the label is added")
-	Eventually(func() error {
-		err := updateClusterLabelByAPI(httpClient, patches, utils.GetClusterClaimID(&cluster))
-		if err != nil {
-			return err
-		}
-		managedCluster, err := getManagedClusterByName(httpClient, cluster.Name)
-		if err != nil {
-			return err
-		}
-		if managedCluster == nil {
-			return fmt.Errorf("no managedcluster found")
-		}
-		if val, ok := managedCluster.Labels[labelKey]; ok {
-			if labelVal == val {
-				return nil
-			}
-		}
-		return fmt.Errorf("failed to add label [%s: %s] to cluster %s, labels: %v", labelKey, labelVal, cluster.Name,
-			cluster.Labels)
-	}, 5*time.Minute, 10*time.Second).ShouldNot(HaveOccurred())
-}
-
-func assertRemoveLabel(cluster clusterv1.ManagedCluster, labelKey, labelVal string) {
-	patches := []patch{
-		{
-			Op:    "remove",
-			Path:  "/metadata/labels/" + labelKey,
-			Value: labelVal,
-		},
-	}
-
-	Eventually(func() error {
-		err := updateClusterLabelByAPI(httpClient, patches, utils.GetClusterClaimID(&cluster))
-		if err != nil {
-			return err
-		}
-		managedCluster, err := getManagedClusterByName(httpClient, cluster.Name)
-		if err != nil {
-			return err
-		}
-		if val, ok := managedCluster.Labels[labelKey]; ok {
-			if val == labelVal {
-				return fmt.Errorf("the label %s:%s should be deleted from cluster %s", labelKey, labelVal, cluster.Name)
-			}
-		}
-		return nil
-	}, 5*time.Minute, 10*time.Second).ShouldNot(HaveOccurred())
-}
-
-func updateClusterLabelByAPI(client *http.Client, patches []patch, managedClusterID string) error {
-	updateLabelUrl := fmt.Sprintf("%s/global-hub-api/v1/managedcluster/%s",
-		testOptions.GlobalHub.Nonk8sApiServer, managedClusterID)
-	// set method and body
-	jsonBody, err := json.Marshal(patches)
-	if err != nil {
-		return err
-	}
-	req, err := http.NewRequest("PATCH", updateLabelUrl, bytes.NewBuffer(jsonBody))
-	if err != nil {
-		return err
-	}
-
-	// add header
-	req.Header.Add("Accept", "application/json")
-
-	// do request
-	response, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err := response.Body.Close(); err != nil {
-			log.Printf("failed to close response body: %v", err)
-		}
-	}()
-	return nil
-}
 
 func updateClusterLabel(managedClusterName, labelStr string) error {
 	leafhubName, _ := strings.CutSuffix(managedClusterName, "-cluster1")
