--- test/manifest/app/app-helloworld-appsub.yaml
@@ -1,74 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: git-application-samples-ns
----
-apiVersion: apps.open-cluster-management.io/v1
-kind: Channel
-metadata:
-  name: git-application-samples
-  namespace: git-application-samples-ns
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-spec:
-  pathname: 'https://github.com/stolostron/application-samples.git'
-  type: Git
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: helloworld
----
-apiVersion: app.k8s.io/v1beta1
-kind: Application
-metadata:
-  name: helloworld
-  namespace: helloworld
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-spec:
-  componentKinds:
-    - group: apps.open-cluster-management.io
-      kind: Subscription
-  descriptor: {}
-  selector:
-    matchExpressions:
-      - key: app
-        operator: In
-        values:
-          - helloworld
----
-apiVersion: apps.open-cluster-management.io/v1
-kind: Subscription
-metadata:
-  annotations:
-    apps.open-cluster-management.io/git-branch: main
-    apps.open-cluster-management.io/git-path: helloworld
-    apps.open-cluster-management.io/reconcile-option: merge
-  labels:
-    app: helloworld
-    app.kubernetes.io/part-of: helloworld
-    apps.open-cluster-management.io/reconcile-rate: medium
-    global-hub.open-cluster-management.io/global-resource: ""
-  name: helloworld-appsub
-  namespace: helloworld
-spec:
-  channel: git-application-samples-ns/git-application-samples
-  placement:
-    placementRef:
-      kind: PlacementRule
-      name: helloworld-placement
----
-apiVersion: apps.open-cluster-management.io/v1
-kind: PlacementRule
-metadata:
-  annotations: {}
-  labels:
-    app: helloworld
-    global-hub.open-cluster-management.io/global-resource: ""
-  name: helloworld-placement
-  namespace: helloworld
-spec:
-  clusterSelector:
-    matchLabels:
-      app: test
