--- test/manifest/crd/0000_00_apps.open-cluster-management.io_subscriptionstatuses_crd_v1alpha1.yaml
@@ -1,70 +1,0 @@
-apiVersion: apiextensions.k8s.io/v1
-kind: CustomResourceDefinition
-metadata:
-  annotations:
-    controller-gen.kubebuilder.io/version: v0.5.0
-  creationTimestamp: null
-  name: subscriptionstatuses.apps.open-cluster-management.io
-spec:
-  group: apps.open-cluster-management.io
-  names:
-    kind: SubscriptionStatus
-    listKind: SubscriptionStatusList
-    plural: subscriptionstatuses
-    shortNames:
-    - appsubstatus
-    singular: subscriptionstatus
-  scope: Namespaced
-  versions:
-  - name: v1alpha1
-    schema:
-      openAPIV3Schema:
-        description: SubscriptionStatus defines the status of package deployments
-        properties:
-          apiVersion:
-            description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
-            type: string
-          kind:
-            description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
-            type: string
-          metadata:
-            type: object
-          statuses:
-            description: Statuses represents all the resources deployed by the subscription per cluster
-            properties:
-              packages:
-                items:
-                  description: SubscriptionUnitStatus defines status of a package deployment.
-                  properties:
-                    apiVersion:
-                      type: string
-                    kind:
-                      type: string
-                    lastUpdateTime:
-                      format: date-time
-                      type: string
-                    message:
-                      type: string
-                    name:
-                      type: string
-                    namespace:
-                      type: string
-                    phase:
-                      description: PackagePhase defines the phasing of a Package
-                      type: string
-                  required:
-                  - lastUpdateTime
-                  type: object
-                type: array
-            type: object
-        type: object
-    served: true
-    storage: true
-    subresources:
-      status: {}
-status:
-  acceptedNames:
-    kind: ""
-    plural: ""
-  conditions: []
-  storedVersions: []
