--- test/manifest/policy/inform-limitrange-policy-placement.yaml
@@ -1,94 +1,0 @@
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: global-placement
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: policy-limitrange
-  namespace: global-placement
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-  annotations:
-    policy.open-cluster-management.io/standards: NIST-CSF
-    policy.open-cluster-management.io/categories: PR.IP Information Protection Processes and Procedures
-    policy.open-cluster-management.io/controls: PR.IP-1 Baseline Configuration
-spec:
-  remediationAction: inform
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: policy-limitrange-container-mem-limit-range
-        spec:
-          remediationAction: inform
-          severity: medium
-          namespaceSelector:
-            exclude:
-              - kube-*
-            include:
-              - default
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: LimitRange
-                metadata:
-                  name: container-mem-limit-range
-                spec:
-                  limits:
-                    - default:
-                        memory: 512Mi
-                      defaultRequest:
-                        memory: 256Mi
-                      type: Container
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: binding-policy-limitrange
-  namespace: global-placement
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-placementRef:
-  name: placement-policy-limitrange
-  kind: Placement
-  apiGroup: cluster.open-cluster-management.io
-subjects:
-- name: policy-limitrange
-  kind: Policy
-  apiGroup: policy.open-cluster-management.io
----
-apiVersion: cluster.open-cluster-management.io/v1beta1
-kind: Placement
-metadata:
-  name: placement-policy-limitrange
-  namespace: global-placement
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-spec:
-  clusterSets:
-    - clusterset1
----
-apiVersion: cluster.open-cluster-management.io/v1beta2
-kind: ManagedClusterSetBinding
-metadata:
-  name: clusterset1
-  namespace: global-placement
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-spec:
-  clusterSet: clusterset1
----
-apiVersion: cluster.open-cluster-management.io/v1beta2
-kind: ManagedClusterSet
-metadata:
-  name: clusterset1
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-spec:
-  clusterSelector:
-    selectorType: ExclusiveClusterSetLabel
