--- test/manifest/policy/inform-limitrange-policy.yaml
@@ -1,76 +1,0 @@
-apiVersion: policy.open-cluster-management.io/v1
-kind: Policy
-metadata:
-  name: policy-limitrange
-  namespace: default
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-  annotations:
-    policy.open-cluster-management.io/standards: NIST-CSF
-    policy.open-cluster-management.io/categories: PR.IP Information Protection Processes and Procedures
-    policy.open-cluster-management.io/controls: PR.IP-1 Baseline Configuration
-spec:
-  remediationAction: inform
-  disabled: false
-  policy-templates:
-    - objectDefinition:
-        apiVersion: policy.open-cluster-management.io/v1
-        kind: ConfigurationPolicy
-        metadata:
-          name: policy-limitrange-container-mem-limit-range
-        spec:
-          remediationAction: inform
-          severity: medium
-          namespaceSelector:
-            exclude:
-              - kube-*
-            include:
-              - default
-          object-templates:
-            - complianceType: musthave
-              objectDefinition:
-                apiVersion: v1
-                kind: LimitRange
-                metadata:
-                  name: container-mem-limit-range
-                spec:
-                  limits:
-                    - default:
-                        memory: 512Mi
-                      defaultRequest:
-                        memory: 256Mi
-                      type: Container
----
-apiVersion: policy.open-cluster-management.io/v1
-kind: PlacementBinding
-metadata:
-  name: binding-policy-limitrange
-  namespace: default
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-placementRef:
-  name: placementrule-policy-limitrange
-  kind: PlacementRule
-  apiGroup: apps.open-cluster-management.io
-subjects:
-  - name: policy-limitrange
-    kind: Policy
-    apiGroup: policy.open-cluster-management.io
----
-apiVersion: apps.open-cluster-management.io/v1
-kind: PlacementRule
-metadata:
-  name: placementrule-policy-limitrange
-  namespace: default
-  labels:
-    global-hub.open-cluster-management.io/global-resource: ""
-spec:
-  clusterConditions:
-    - status: 'True'
-      type: ManagedClusterConditionAvailable
-  clusterSelector:
-    matchExpressions:
-      - key: global-policy
-        operator: In
-        values:
-          - test
